redesign version selector menu [WIP]
