+++ app/components/header/version_selector_component.html.erb
@@ -1,0 +1,29 @@
+
+<div data-controller="dropdown" class="relative">
+  <button type="button" data-action="dropdown#toggle click@window->dropdown#hide">
+    <span class="text-lg">v<%= @current_version %></span>
+    <svg xmlns="http://www.w3.org/2000/svg" class="inline h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
+      <path fill-rule="evenodd" d="M5.293 7.293a1 1 0 011.414 0L10 10.586l3.293-3.293a1 1 0 111.414 1.414l-4 4a1 1 0 01-1.414 0l-4-4a1 1 0 010-1.414z" clip-rule="evenodd"></path>
+    </svg>
+  </button>
+  <div
+    data-dropdown-target="menu"
+    class="hidden transition transform origin-top-right absolute right-0 mt-2 w-32 rounded-md shadow-lg bg-white overflow-hidden ring-1 ring-black ring-opacity-5 focus:outline-none" role="menu" aria-orientation="vertical" aria-labelledby="user-menu-button" tabindex="-1"
+    data-transition-enter-from="opacity-0 scale-95"
+    data-transition-enter-to="opacity-100 scale-100"
+    data-transition-leave-from="opacity-100 scale-100"
+    data-transition-leave-to="opacity-0 scale-95">
+    <% @ruby_versions.each do |v| %>
+      <a href="<%= route_for_version(v) %>" class="flex justify-between items-center text-left space-x-1 px-4 py-2 text-gray-700 hover:bg-gray-200 dark:hover:bg-gray-800 <%= "text-blue-800 bg-blue-200" if v.version == @current_version %>" role="menuitem" tabindex="-1">
+        <span class="text-l font-semibold"><%= v.version %></span>
+        <span>
+          <% if v.version == @current_version %>
+            <svg xmlns="http://www.w3.org/2000/svg" class="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
+              <path fill-rule="evenodd" d="M16.707 5.293a1 1 0 010 1.414l-8 8a1 1 0 01-1.414 0l-4-4a1 1 0 011.414-1.414L8 12.586l7.293-7.293a1 1 0 011.414 0z" clip-rule="evenodd" />
+            </svg>
+          <% end %>
+        </span>
+      </a>
+    <% end %>
+  </div>
+</div>
