+++ app/components/header_component.html.erb
@@ -18,7 +18,7 @@
           </button>
         </div>
         <div class="hidden md:flex md:space-x-4">
-          <%= render 'layouts/version_selector' %>
+          <%= render Header::VersionSelectorComponent.new(current_version: "3.1", versions: RubyConfig.ruby_versions) %>
           <%= render 'layouts/theme_selector' %>
         </div>
       </div>
@@ -35,7 +35,7 @@
         </div>
         <div class="flex justify-between space-x-4 items-center">
           <div class="flex justify-between items-center space-x-4">
-            <%= render 'layouts/version_selector' %>
+            <%= render Header::VersionSelectorComponent.new(current_version: "3.1", versions: RubyConfig.ruby_versions) %>
             <%= render 'layouts/theme_selector' %>
           </div>
           <button data-action="click->header#toggleMobileMenu" class="group px-2 py-1 -my-1 border border-solid border-gray-300 dark:border-gray-500 text-xl curser-pointer rounded-lg inline-flex items-center justify-center focus:outline-none hover:bg-gray-200 dark:hover:bg-gray-500">
